Respond to review:
> The entry Typing Rules in the table of contents displays differently
> (bolder) than the others.
I removed bolding I had accidentally added.

> Incidentally, I had not put these by-chapter TOC in the chapters I
> converted. Probably we want them, in which case we should add that as
> an issue.
I see that @psteckler has added this to issue #20 already.

> In the typing rules, for the construct foo \in Bar, there's no space
> around the in, which looks not so nice.
I do not understand this comment. I checked the TeX version and there
is no explicit extra spacing around the \in operator.
I'm seeing the usual spacing around math operators for \in
in the html rendered version. So I made no changes for this.

> For the mutual inductive declarations, in the TeX version, I see Ind
> [1] where in some of the Sphinx ones, I see Ind [].
The automatic translation of these Ind examples was all messed up,
so I had to reconstruct them from scratch (actually with help from
the .tex sources). But I missed the [1] on one of them.
I fixed it.

> There's a TODO after TreeExample. Is that meant to stay in?
This is an issue that goes beyond conversion to Sphinx. It seems
that they commentary following does not match with the preceding
example. This will require more editorial work. I left the TODO
in and expanded on the issue some. This can be addressed after the
conversion is done and we start doing edits to improve the manual.

> In the section about impredicative Set, there's a sentence about the
> "ordinary coqtop command". In the chapter "The Coq commands", I put
> all the names like coqtop, coqc, etc., enclosed in single
> backticks. If that's a good idea (?), we should do that uniformly.
I changed the two instances of coqtop there to use `coqdoc`.